Rename validation handlers to validator creators
