+++ src/schemaType/BaseScalarType.ts
@@ -2,7 +2,7 @@ import { cloneDeep, set as setIn, toPath } from 'lodash';
 import { InvalidParameterError } from '../errors';
 import type { DecryptFunc, EncryptFunc, GenericObject, MvRecord } from '../types';
 import { getFromMvArray } from '../utils';
-import { handleRequiredValidation } from '../validators';
+import { createRequiredValidator } from '../validators';
 import BaseSchemaType from './BaseSchemaType';
 import type { SchemaTypeDefinitionBoolean } from './BooleanType';
 import type { SchemaTypeDefinitionISOCalendarDateTime } from './ISOCalendarDateTimeType';
@@ -107,7 +107,7 @@ abstract class BaseScalarType extends BaseSchemaType {
 		return (
 			await Promise.all(
 				this.validators
-					.concat(handleRequiredValidation(this.required, this.validateRequired))
+					.concat(createRequiredValidator(this.required, this.validateRequired))
 					.map(async ({ validator, message }) => {
 						const isValid = await validator(value, document);
 						return isValid ? ISVALID_SYMBOL : message;
