+++ src/validators/__tests__/createRequiredValidator.test.ts
@@ -1,16 +1,16 @@
-import handleRequiredValidation from '../handleRequiredValidation';
+import createRequiredValidator from '../createRequiredValidator';
 
 const defaultValidator = (): false => false;
 
 test('should return validator function which evaluates to true and empty message if required is false', (): void => {
-	const { validator, message } = handleRequiredValidation(false, defaultValidator);
+	const { validator, message } = createRequiredValidator(false, defaultValidator);
 	expect(validator).toBeInstanceOf(Function);
 	expect(validator(null, { foo: null })).toBe(true);
 	expect(message).toBe('Property is required');
 });
 
 test('should return default validator and message if required is true', (): void => {
-	const { validator, message } = handleRequiredValidation(true, defaultValidator);
+	const { validator, message } = createRequiredValidator(true, defaultValidator);
 	expect(validator).toEqual(defaultValidator);
 	expect(message).toBe('Property is required');
 });
