+++ src/validators/__tests__/createTypeValidator.test.ts
@@ -1,9 +1,9 @@
-import handleTypeValidation from '../handleTypeValidation';
+import createTypeValidator from '../createTypeValidator';
 
 const defaultValidator = (): false => false;
 
 test('should return default validator and message', (): void => {
-	const { validator, message } = handleTypeValidation(defaultValidator);
+	const { validator, message } = createTypeValidator(defaultValidator);
 	expect(validator).toEqual(defaultValidator);
 	expect(message).toBe('Property cannot be cast into the defined type');
 });
